requirements.txt
streamlit==1.36.0
pandas==2.1.4
openpyxl==3.1.2
transformers==4.38.1
sentence-transformers==2.6.1
faiss-cpu==1.7.4
python-dotenv==1.0.1
torch==2.1.2
numpy==1.26.4
scikit-learn==1.4.1
tqdm==4.66.2
accelerate==0.27.2